# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  post_handler.CreatePostRequest:
    properties:
      content:
        type: string
      media_items:
        items:
          $ref: '#/definitions/post_handler.MediaItemInput'
        maxItems: 9
        type: array
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - title
    type: object
  post_handler.CreatePostResponse:
    properties:
      author_avatar_url:
        type: string
      author_bio:
        type: string
      author_email:
        type: string
      author_full_name:
        type: string
      author_id:
        type: integer
      author_username:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      media:
        items:
          $ref: '#/definitions/post_handler.PostMediaResponse'
        type: array
      tags:
        items:
          $ref: '#/definitions/post_handler.TagResponse'
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  post_handler.GetPostMedia:
    properties:
      id:
        type: integer
      position:
        type: integer
      type:
        type: string
      url:
        type: string
    type: object
  post_handler.GetPostResponse:
    properties:
      author:
        $ref: '#/definitions/post_handler.GetPostUser'
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      media:
        items:
          $ref: '#/definitions/post_handler.GetPostMedia'
        type: array
      tags:
        items:
          $ref: '#/definitions/post_handler.GetPostTag'
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  post_handler.GetPostTag:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  post_handler.GetPostUser:
    properties:
      avatar_url:
        type: string
      full_name:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
...
  post_handler.MediaItemInput:
    properties:
      position:
        maximum: 100
        minimum: 0
        type: integer
      type:
        enum:
        - image
        - video
        type: string
      url:
        maxLength: 512
        type: string
    required:
    - type
    - url
    type: object
  post_handler.PostMediaResponse:
    properties:
      id:
        type: integer
      position:
        type: integer
      type:
        type: string
      url:
        type: string
    type: object
  post_handler.TagResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  post_handler.UpdatePostAuthor:
    properties:
      avatar_url:
        type: string
      full_name:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  post_handler.UpdatePostRequest:
    properties:
      content:
        type: string
      media_items:
        items:
          $ref: '#/definitions/post_handler.MediaItemInput'
        maxItems: 9
        type: array
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 255
        minLength: 1
        type: string
    type: object
  post_handler.UpdatePostResponse:
    properties:
      author:
        $ref: '#/definitions/post_handler.UpdatePostAuthor'
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      media:
        items:
          $ref: '#/definitions/post_handler.PostMediaResponse'
        type: array
      tags:
        items:
          $ref: '#/definitions/post_handler.TagResponse'
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Gateway for Pinstack social media platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pinstack API Gateway
  version: "1.0"
paths:
...
  /posts:
    delete:
      description: Delete an existing post by ID
      parameters:
      - description: Post ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a post
      tags:
      - posts
    get:
      description: Get detailed information about a specific post
      parameters:
      - description: Post ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post information
          schema:
            $ref: '#/definitions/post_handler.GetPostResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get post by ID
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create a new post with title, content, tags and media
      parameters:
      - description: Post creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/post_handler.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Post created successfully
          schema:
            $ref: '#/definitions/post_handler.CreatePostResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new post
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Update an existing post with new data
      parameters:
      - description: Post ID
        in: query
        name: id
        required: true
        type: string
      - description: Post update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/post_handler.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Post updated successfully
          schema:
            $ref: '#/definitions/post_handler.UpdatePostResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a post
      tags:
      - posts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"